+++ src/extension.js
@@ -9,7 +9,27 @@ function activate(context) {
         commands.pullTickets
     );
 
-    context.subscriptions.push(pullTickets);
+    let generateCode = vscode.commands.registerCommand(
+        'extension.generateCode',
+        commands.generateCode
+    );
+
+    let updateTicket = vscode.commands.registerCommand(
+        'extension.updateTicket',
+        commands.updateTicket
+    );
+
+    let createWorkItems = vscode.commands.registerCommand(
+        'extension.createWorkItems',
+        commands.createWorkItems
+    );
+
+    context.subscriptions.push(
+        pullTickets,
+        generateCode,
+        updateTicket,
+        createWorkItems
+    );
 }
 
 function deactivate() {}